feat: parent tree tests - failing - wip
issue #368

+++ backend/src/main/java/ca/bc/gov/backendstartapi/provider/Provider.java
@@ -53,7 +53,7 @@ public interface Provider {
     return List.of();
   }
 
-  default List<ParentTreeLatLongDto> getParentTreeLatLongByIdList(List<Integer> ptIds) {
+  default List<ParentTreeLatLongDto> getParentTreeLatLongByIdList(List<Long> ptIds) {
     return List.of();
   }
